+++ src/pages/mypage/pages/purchase-list/index.style.ts
